Remove redundant i128 same-type casts in adjust params
clippy 0.1.96 (CI's stable) flags `margin_delta as i128` as
unnecessary_cast — scale_to_6dec already returns i128. Drop the casts in
adjust_taker / adjust_maker. (Local clippy 0.1.91 didn't catch it.)

+++ src/client/trades.rs
@@ -181,7 +181,7 @@ impl PerpClient {
 
         let wire_params = crate::contracts::AdjustTakerParams {
             posId: params.pos_id,
-            marginDelta: margin_delta as i128,
+            marginDelta: margin_delta,
             perpDelta: perp_delta,
             amt1Limit: U256::from(params.amt1_limit),
         };
@@ -219,7 +219,7 @@ impl PerpClient {
 
         let wire_params = crate::contracts::AdjustMakerParams {
             posId: params.pos_id,
-            marginDelta: margin_delta as i128,
+            marginDelta: margin_delta,
             liquidityDelta: params.liquidity_delta,
             amt0Limit: U256::from(params.amt0_limit),
             amt1Limit: U256::from(params.amt1_limit),
